Move Bmad-X utils to new utils directory
