+++ cheetah/accelerator/quadrupole.py
@@ -7,7 +7,10 @@ from matplotlib.patches import Rectangle
 from scipy.constants import physical_constants
 from torch import Size, nn
 
-from cheetah.bmadx_utils import (
+from cheetah.particles import Beam, ParameterBeam, ParticleBeam
+from cheetah.track_methods import base_rmatrix, misalignment_matrix
+from cheetah.utils import UniqueNameGenerator
+from cheetah.utils.bmadx import (
     bmad_to_cheetah_coords,
     calculate_quadrupole_coefficients,
     cheetah_to_bmad_coords,
@@ -15,9 +18,6 @@ from cheetah.bmadx_utils import (
     offset_particle_set,
     offset_particle_unset,
 )
-from cheetah.particles import Beam, ParameterBeam, ParticleBeam
-from cheetah.track_methods import base_rmatrix, misalignment_matrix
-from cheetah.utils import UniqueNameGenerator
 
 from .element import Element
